refactor(prometheus): clean up collector package layout
