+++ src/go/plugin/go.d/collector/prometheus/promprofiles/default_catalog.go
@@ -1,0 +1,40 @@
+// SPDX-License-Identifier: GPL-3.0-or-later
+
+package promprofiles
+
+import (
+	"sync"
+
+	"github.com/netdata/netdata/go/plugins/pkg/executable"
+)
+
+var (
+	defaultCatalogMu           sync.Mutex
+	defaultCatalog             Catalog
+	defaultCatalogLoaded       bool
+	defaultCatalogLoader       = loadFromDefaultDirs
+	defaultCatalogCacheEnabled = func() bool { return executable.Name != "test" }
+)
+
+func DefaultCatalog() (Catalog, error) {
+	if !defaultCatalogCacheEnabled() {
+		return defaultCatalogLoader()
+	}
+
+	defaultCatalogMu.Lock()
+	defer defaultCatalogMu.Unlock()
+
+	if defaultCatalogLoaded {
+		return defaultCatalog, nil
+	}
+
+	catalog, err := defaultCatalogLoader()
+	if err != nil {
+		return Catalog{}, err
+	}
+
+	defaultCatalog = catalog
+	defaultCatalogLoaded = true
+
+	return defaultCatalog, nil
+}
